tvforms: fix TDataRec padding issue
This makes resource files ABI-compatible with Borland C++.

+++ examples/tvforms/genparts.h
@@ -61,7 +61,7 @@ const int
 
 struct TDescrRec 
 {
-    ushort textLen;
+    uint textLen;
     char textData[descrLen];
 };
